requirements.txt
aniso8601==1.2.0
asn1crypto==0.22.0
BeautifulSoup==3.2.1
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
click==6.7
cryptography==1.9
DateTime==4.2
enum34==1.1.6
Flask==0.12.1
Flask-Ask==0.9.4
Flask-SQLAlchemy==2.2
gunicorn==19.7.1
idna==2.5
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
psycopg2==2.7.1
pycparser==2.17
PyJWT==1.5.3
pyOpenSSL==17.0.0
python-dateutil==2.6.0
python-firebase==1.2
pytz==2017.2
PyYAML==3.12
requests==2.18.1
satori-sdk-python==1.0.3
six==1.10.0
SQLAlchemy==1.1.10
twilio==6.8.2
urllib3==1.21.1
Werkzeug==0.12.2
zope.interface==4.4.2

